Fixing rocm gptq by using triton code too (renamed cuda into triton).
